test: 신규 소스 단위시험을 tests/cases 로 이관한다 (#5140, #5174)
`scripts/rust-unit-test-tiers.mjs --check` 는 제품 소스 `#[cfg(test)]` 의 테스트 수
**증가**를 실패시킨다(신규는 `tests/cases/` 통합 테스트로만). 이 브랜치는 그 게이트를
어기고 있었다:

    source unit test 증가 금지: src/parser/tags.rs::tests#1 (9 > 4)
    source unit test 증가 금지: src/serializer/body_text.rs::tests#1 (35 > 34)
    source unit test 증가 금지: src/serializer/hwpx/section.rs::tests#1 (96 > 93)
    source-side test 총량 증가 금지: 4234 > 4225

- `src/parser/tags.rs` 의 #5140 사상표 시험 5개는 검사 대상이 모두 공개 항목이라
  `tests/cases/issue_5140_hancom_symbol_plane15.rs` 로 그대로 옮겼다.
- #5174 가 늘린 4개는 `render_hp_t_content`·`compute_control_mask` 가 `pub(crate)` 라
  그대로 옮길 수 없어, 앞 커밋의 `tests/cases/issue_5174_nbspace_representation.rs` 가
  같은 계약을 **문서 왕복 축**으로 다시 쓴다. 남긴 단위시험은 두 출처를 한 번에 도는
  형태로 합쳐 기준선(34·93)을 지킨다.

`tests/suites/unit-test-tiers.json` 은 `--generate` 재생성분이고 총량은 기준선인
4225 로 돌아온다. `tests/generated/`·`tests/suites/manifest.json` 은 .gitignore 대상이라
커밋에 포함되지 않는다.
