+++ tests/suites/unit-test-tiers.json
@@ -763,7 +763,7 @@
       "id": "src/document_core/commands/document.rs::set_document_tests#1",
       "file": "src/document_core/commands/document.rs",
       "sourcePath": "src/document_core/commands/document.rs",
-      "line": 3087,
+      "line": 3100,
       "module": "set_document_tests",
       "testAttributes": 2,
       "tier": "white_box",
@@ -776,7 +776,7 @@
       "id": "src/document_core/commands/document.rs::validate_linesegs_tests#1",
       "file": "src/document_core/commands/document.rs",
       "sourcePath": "src/document_core/commands/document.rs",
-      "line": 2276,
+      "line": 2289,
       "module": "validate_linesegs_tests",
       "testAttributes": 25,
       "tier": "white_box",
@@ -2076,7 +2076,7 @@
       "id": "src/parser/body_text.rs::tests#1",
       "file": "src/parser/body_text.rs",
       "sourcePath": "src/parser/body_text/tests.rs",
-      "line": 1162,
+      "line": 1179,
       "module": "tests",
       "testAttributes": 33,
       "tier": "white_box",
@@ -2349,7 +2349,7 @@
       "id": "src/parser/hwpx/section.rs::tests#1",
       "file": "src/parser/hwpx/section.rs",
       "sourcePath": "src/parser/hwpx/section.rs",
-      "line": 7203,
+      "line": 7218,
       "module": "tests",
       "testAttributes": 94,
       "tier": "white_box",
@@ -2427,7 +2427,7 @@
       "id": "src/parser/tags.rs::tests#1",
       "file": "src/parser/tags.rs",
       "sourcePath": "src/parser/tags.rs",
-      "line": 390,
+      "line": 463,
       "module": "tests",
       "testAttributes": 4,
       "tier": "white_box",
@@ -3389,7 +3389,7 @@
       "id": "src/serializer/body_text.rs::tests#1",
       "file": "src/serializer/body_text.rs",
       "sourcePath": "src/serializer/body_text.rs",
-      "line": 1273,
+      "line": 1294,
       "module": "tests",
       "testAttributes": 34,
       "tier": "white_box",
@@ -3532,7 +3532,7 @@
       "id": "src/serializer/hwpx/field.rs::tests#1",
       "file": "src/serializer/hwpx/field.rs",
       "sourcePath": "src/serializer/hwpx/field.rs",
-      "line": 282,
+      "line": 290,
       "module": "tests",
       "testAttributes": 17,
       "tier": "white_box",
@@ -3649,7 +3649,7 @@
       "id": "src/serializer/hwpx/section.rs::tests#1",
       "file": "src/serializer/hwpx/section.rs",
       "sourcePath": "src/serializer/hwpx/section.rs",
-      "line": 3167,
+      "line": 3212,
       "module": "tests",
       "testAttributes": 93,
       "tier": "white_box",
@@ -4104,7 +4104,7 @@
     {
       "id": "src/serializer/body_text.rs::cfg-fn#1",
       "file": "src/serializer/body_text.rs",
-      "line": 530,
+      "line": 535,
       "kind": "fn",
       "testAttributes": 0
     },
